Sequence Styles
To understand sequence styles, it is important to understand collections. The concept of
collections and sequence styles work in parallel. The collection in YAML is represented with
proper sequence styles. If you want to refer proper sequencing of tags, always refer to
collections. Collections in YAML are indexed by sequential integers starting with zero as
represented in arrays. The focus of sequence styles begins with collections.